refactor: consolidate tsm-embedder into tsmd for 2-binary architecture
Merge tsm-embedder into tsmd and extract watcher from thread to child
process, reducing 3 binaries (tsm, tsmd, tsm-embedder) to 2 (tsm, tsmd).
This eliminates macOS Gatekeeper friction and aligns with docker/dockerd
pattern.

- Add --embedder, --fs-watcher, --model, --no-idle-timeout flags to tsmd
- Split src/bin/tsmd.rs (771 lines) into src/bin/tsmd/ with 6 modules
- Move embedder daemon code from src/embedder.rs to embedder_mode.rs
- Keep src/embedder.rs as pure inference library
- Watcher uses daemon.sock IPC instead of direct DB access
- Watcher reload via SIGHUP instead of mpsc channel
- Extract spawn_child() helper to prevent orphaned processes
- Add ADR-0005 documenting the design decision

Closes #82
